4/27 HomeTop.java 시간 표시를 기획서에 있는것처럼 "xx:xx"로 수정 2

+++ PM01/src/com/aga/mine/mains/HomeTop.java
@@ -223,7 +223,7 @@ public class HomeTop extends CCLayer{
 		long min = (secondOfInt % (60 * 60)) / 60;
 		long sec = secondOfInt % 60;
 		
-		String deadlineText = String.format("%02d:%02d", min, sec);
+		String deadlineText = String.format("%02d:%02d", (int)min, (int)sec);
 
 		return deadlineText;
 	}
